feat: создание слоя рек, где объекты - сами реки, а не сегменты + фильтрация
