+++ src/river/point_selection_tool.py
