+++ src/river/river.py
@@ -1,39 +1,35 @@
 import os
+from typing import List, Optional
 import processing
 from qgis.core import (
     QgsProject,
     QgsCoordinateReferenceSystem,
     QgsCoordinateTransform,
     QgsPointXY,
-    QgsRasterLayer,
     QgsVectorLayer,
     QgsGeometry,
     QgsField,
     QgsRaster,
     QgsFeature,
-    QgsApplication,
 )
 from qgis.PyQt.QtCore import QVariant, QEventLoop
 from qgis.PyQt.QtWidgets import (
     QMessageBox,
     QInputDialog
 )
-from processing_saga_nextgen.saga_nextgen_plugin import SagaNextGenAlgorithmProvider
 from qgis.utils import iface
-from .common import *
 from .point_selection_tool import PointSelectionTool
+from .layers.rivers_by_object_filtered import build_rivers_by_object_filtered
+from ..river.layers.basins import build_basins_layer
+from ..river.layers.rivers_and_points import build_rivers_and_points_layer
+from ..river.layers.rivers_merged import build_merged_layer
+from ..river.layers.utils import load_quickosm_layer
+from ..common import *
 
 RIVER_FILTERS = {
     "strahler_order": (">=", 2),
     "length": (">", 1000),
 }
-
-def load_saga_algorithms():
-    provider = SagaNextGenAlgorithmProvider()
-    provider.loadAlgorithms()
-
-    QgsApplication.processingRegistry().addProvider(provider=provider)
-
 
 def transform_bbox(x_min, x_max, y_min, y_max, from_epsg, to_epsg):
     # Создаем объекты систем координат
@@ -71,61 +67,16 @@ def river(project_folder):
     dem_layer = add_dem_layer(dem_path)
     reprojected_relief = reproject_dem(dem_path)
 
-    load_saga_algorithms()
-    # Использовать SAGA Fill Sinks для извлечения водосборов
+    # Добавить заполненные области водосбора в проект
     basins_path = os.path.join(project_folder, "basins.sdat")
-    processing.run(
-        "sagang:fillsinkswangliu",
-        {
-            "ELEV": reprojected_relief,
-            "FILLED": "TEMPORARY_OUTPUT",
-            "FDIR": "TEMPORARY_OUTPUT",
-            "WSHED": basins_path,
-            "MINSLOPE": 0.01,
-        },
-    )
-
-    # Сохранить и добавить заполненные области водосбора в проект
-    basins = QgsRasterLayer(basins_path, "basins")
+    basins = build_basins_layer(reprojected_relief, basins_path)
     QgsProject.instance().addMapLayer(basins)
 
+    # Добавить объединенный слой рек и ручьев в проект
     # Использовать QuickOSM для запроса данных о водных путях на заданной территории
     extent = transform_bbox(bbox[0], bbox[2], bbox[1], bbox[3], 4326, 3857)
-
-    # Загрузить реки
-    rivers_layer = load_quickosm_layer("rivers", "waterway", "river", extent)
-    # Загрузить ручьи
-    streams_layer = load_quickosm_layer("streams", "waterway", "stream", extent)
-
-    # Объединить слои рек и ручьев
-    merged_filepath = os.path.join(project_folder, "merge_result.gpkg")
-    merge_result = processing.run(
-        "qgis:mergevectorlayers",
-        {
-            "LAYERS": [rivers_layer, streams_layer],
-            "CRS": rivers_layer.crs(),
-            "OUTPUT": "TEMPORARY_OUTPUT",
-        },
-    )["OUTPUT"]
-    merge_result = processing.run(
-        "native:dissolve",
-        {
-            "INPUT": merge_result,
-            "FIELD": [],
-            "SEPARATE_DISJOINT": False,
-            "OUTPUT": "TEMPORARY_OUTPUT",
-        },
-    )["OUTPUT"]
-    processing.run(
-        "native:multiparttosingleparts",
-        {
-            "INPUT": merge_result,
-            "OUTPUT": merged_filepath,
-        },
-    )["OUTPUT"]
-
-    # Добавить объединенный слой в проект
-    rivers_merged = QgsVectorLayer(merged_filepath, "rivers_merged", "ogr")
+    merged_path = os.path.join(project_folder, "merge_result.gpkg")
+    rivers_merged = build_merged_layer(extent, merged_path)
     QgsProject.instance().addMapLayer(rivers_merged)
 
     # Загрузить полигональные данные о водных объектах
@@ -143,7 +94,7 @@ def river(project_folder):
     start_x = processing.run(
         "native:fieldcalculator",
         {
-            "INPUT": merged_filepath,
+            "INPUT": merged_path,
             "FIELD_NAME": "start_x",
             "FIELD_TYPE": 0,
             "FIELD_LENGTH": 0,
@@ -240,41 +191,18 @@ def river(project_folder):
     line_provider.changeAttributeValues(changes)
     end_y.commitChanges()
 
-    length_result = processing.run(
-        "native:fieldcalculator",
-        {
-            "INPUT": end_y,
-            "FIELD_NAME": "length",
-            "FIELD_TYPE": 0,
-            "FIELD_LENGTH": 10,
-            "FIELD_PRECISION": 3,
-            "FORMULA": "$length",
-            "OUTPUT": "TEMPORARY_OUTPUT",
-        },
-    )["OUTPUT"]
-    if isinstance(length_result, QgsVectorLayer):
-        rivers_with_length = length_result
-    else:
-        rivers_with_length = QgsVectorLayer(length_result, "rivers_with_length", "ogr")
-    rivers_with_order = compute_strahler(rivers_with_length)
-    rivers_filtered = filter_rivers_by_params(rivers_with_order, RIVER_FILTERS)
-    QgsProject.instance().addMapLayer(rivers_filtered)
+    # Добавим слой рек, где объекты - сами реки, а не сегменты.
+    rivers_by_object_filtered_path = os.path.join(
+        project_folder, "rivers_by_object_filtered.gpkg"
+    )
+    rivers_by_object_filtered = build_rivers_by_object_filtered(
+        end_y, RIVER_FILTERS, rivers_by_object_filtered_path
+    )
+    QgsProject.instance().addMapLayer(rivers_by_object_filtered)
 
     # Определить максимальную высоту для каждой линии
-    max_z = processing.run(
-        "native:fieldcalculator",
-        {
-            "INPUT": end_y,
-            "FIELD_NAME": "max_z",
-            "FIELD_TYPE": 0,
-            "FIELD_LENGTH": 0,
-            "FIELD_PRECISION": 0,
-            "FORMULA": 'if("start_z" > "end_z", "start_z", "end_z")',
-            "OUTPUT": f"{project_folder}rivers_with_points.gpkg",
-        },
-    )["OUTPUT"]
-
-    rivers_and_points = QgsVectorLayer(max_z, "rivers_and_points", "ogr")
+    rivers_and_points_path = os.path.join(project_folder, "rivers_with_points.gpkg")
+    rivers_and_points = build_rivers_and_points_layer(end_y, rivers_and_points_path)
     QgsProject.instance().addMapLayer(rivers_and_points)
 
     # Создать слой точек максимальной высоты
@@ -346,122 +274,7 @@ def river(project_folder):
     QgsProject.instance().addMapLayer(point_layer)
 
 
-def load_quickosm_layer(
-    layer_name,
-    key,
-    value,
-    extent,
-    output_path="TEMPORARY_OUTPUT",
-    quickosm_layername="lines",
-):
-    query = processing.run(
-        "quickosm:buildqueryextent",
-        {
-            "KEY": key,
-            "VALUE": value,
-            "EXTENT": extent,
-            "TIMEOUT": 25,
-        },
-    )
-    download_result = processing.run(
-        "native:filedownloader",
-        {
-            "URL": query["OUTPUT_URL"],
-            "OUTPUT": output_path,
-        },
-    )["OUTPUT"]
-    layer = iface.addVectorLayer(
-        download_result + f"|layername={quickosm_layername}", layer_name, "ogr"
-    )
-    return layer
-
-
-def compute_strahler(rivers_layer):
-    node_map = {}
-    edges = []
-    next_node = 0
-
-    for feat in rivers_layer.getFeatures():
-        sx, sy, sz = feat["start_x"], feat["start_y"], feat["start_z"]
-        ex, ey, ez = feat["end_x"], feat["end_y"], feat["end_z"]
-
-        if sz is None or ez is None:
-            continue
-
-        if sz < ez:
-            sx, sy, sz, ex, ey, ez = ex, ey, ez, sx, sy, sz
-
-        for x, y in ((sx, sy), (ex, ey)):
-            if (x, y) not in node_map:
-                node_map[(x, y)] = next_node
-                next_node += 1
-
-        n1 = node_map[(sx, sy)]
-        n2 = node_map[(ex, ey)]
-        edges.append((n1, n2, feat.id()))
-
-    incoming = {i: [] for i in range(next_node)}
-    outgoing = {i: [] for i in range(next_node)}
-    for u, v, fid in edges:
-        outgoing[u].append((v, fid))
-        incoming[v].append((u, fid))
-
-    orders = {}
-
-    def calc_edge(fid, from_node):
-        if fid in orders:
-            return orders[fid]
-        preds = incoming[from_node]
-        if not preds:
-            orders[fid] = 1
-            return 1
-        child = [calc_edge(up_fid, up_node) for up_node, up_fid in preds]
-        m = max(child)
-        orders[fid] = m + 1 if child.count(m) > 1 else m
-        return orders[fid]
-
-    for node in range(next_node):
-        if not outgoing[node]:
-            for _, fid in incoming[node]:
-                calc_edge(fid, node)
-
-    rivers_layer.startEditing()
-    if "strahler_order" not in [f.name() for f in rivers_layer.fields()]:
-        rivers_layer.dataProvider().addAttributes(
-            [QgsField("strahler_order", QVariant.Int)]
-        )
-        rivers_layer.updateFields()
-    idx = rivers_layer.fields().indexOf("strahler_order")
-    for _, _, fid in edges:
-        rivers_layer.changeAttributeValue(fid, idx, orders.get(fid, 0))
-    rivers_layer.commitChanges()
-
-    return rivers_layer
-
-
-def filter_rivers_by_params(rivers_layer, filters, layer_name="rivers_filtered"):
-    expr_parts = []
-    for fld, (op, val) in filters.items():
-        if isinstance(val, str):
-            safe = val.replace("'", "''")
-            val_repr = f"'{safe}'"
-        else:
-            val_repr = str(val)
-        expr_parts.append(f'"{fld}" {op} {val_repr}')
-    expression = " AND ".join(expr_parts) or "TRUE"
-
-    result_path = processing.run(
-        "native:extractbyexpression",
-        {"INPUT": rivers_layer, "EXPRESSION": expression, "OUTPUT": "TEMPORARY_OUTPUT"},
-    )["OUTPUT"]
-    if isinstance(result_path, QgsVectorLayer):
-        filtered = result_path
-    else:
-        filtered = QgsVectorLayer(result_path, layer_name, "ogr")
-    filtered.setName(layer_name)
-    return filtered
-
-def select_analysis_bbox() -> list | None:
+def select_analysis_bbox() -> Optional[List[float]]:
     method, ok = QInputDialog.getItem(
         None,
         "Выбор метода",
